Created function listeners for opening menu navigation button on mobile view, moved functions from main.js to separate files and imported them to main, updated all sites with new nav hatml
