+++ index.html
@@ -42,10 +42,11 @@
       </div>
     </header>
     <nav class="sec">
-      <a class="mobile-nav-menu">
+      <button class="mobile-nav-menu">
         Meny <span class="arrow-top" id="menu-span">&#9662;</span>
-      </a>
-      <ul class="nav sec-cont">
+      </button>
+      <div class="nav sec-cont">
+      <ul id="nav-id">
         <li class="movies-nav">
           <ul class="movies-nav-list">
             <li class="movies-nav-element">
@@ -106,6 +107,7 @@
           </ul>
         </li>
       </ul>
+    </div>
     </nav>
 
     <main class="sec clear-main">
