+++ restaurant.html
@@ -34,7 +34,11 @@
       </div>
     </header>
     <nav class="sec">
-      <ul class="nav sec-cont">
+      <button class="mobile-nav-menu">
+        Meny <span class="arrow-top" id="menu-span">&#9662;</span>
+      </button>
+      <div class="nav sec-cont">
+      <ul id="nav-id">
         <li class="movies-nav">
           <ul class="movies-nav-list">
             <li class="movies-nav-element">
@@ -60,29 +64,42 @@
         <li class="movies-nav">
           <ul class="movies-nav-list">
             <li class="movies-nav-element">
-              <a id="restaurant-id" href="restaurant.html"
+              <a href="restaurant.html" id="restaurant-id"
                 >Restaurang <span class="arrow-top">&#9662;</span></a
               >
             </li>
-            <li class="movies-nav-element"><a href="under-construction.html">Boka bord</a></li>
-            <li class="movies-nav-element"><a href="under-construction.html">Vår menu</a></li>
+            <li class="movies-nav-element">
+              <a href="under-construction.html">Boka bord</a>
+            </li>
+            <li class="movies-nav-element">
+              <a href="under-construction.html">Vår menu</a>
+            </li>
           </ul>
         </li>
         <li class="movies-nav">
           <ul class="movies-nav-list">
             <li class="movies-nav-element">
-              <a href="under-construction.html">Events <span class="arrow-top">&#9662;</span></a>
+              <a href="under-construction.html"
+                >Events <span class="arrow-top">&#9662;</span></a
+              >
             </li>
-            <li class="movies-nav-element"><a href="under-construction.html">Live på bio</a></li>
-            <li class="movies-nav-element"><a href="under-construction.html">Specialvisningar</a></li>
+            <li class="movies-nav-element">
+              <a href="under-construction.html">Live på bio</a>
+            </li>
+            <li class="movies-nav-element">
+              <a href="under-construction.html">Specialvisningar</a>
+            </li>
           </ul>
         </li>
         <li class="movies-nav">
           <ul class="movies-nav-list">
-            <li class="movies-nav-element"><a href="under-construction.html">Kontakt</a></li>
+            <li class="movies-nav-element">
+              <a href="under-construction.html">Kontakt</a>
+            </li>
           </ul>
         </li>
       </ul>
+    </div>
     </nav>
     <main class="sec clear-main">
         <div class="sec-cont">
