+++ src/sections/nav-menu.scss
@@ -1,6 +1,11 @@
 // Nav meny
 .nav {
+  display: block; 
   width: 100%;
+  margin: 0 auto;
+  //padding: 0;
+  ul {
+    
   display: flex;
   flex-direction: row;
 
@@ -48,38 +53,62 @@
       display: none;
     }
   }
-}
+}}
 
 .mobile-nav-menu {
   display: none;
 }
 
 @media screen and (max-width: $breaking-point) {
-  .nav {
-    width: 100%;
-    display: flex;
-    flex-direction: column;
+            .mobile-nav-menu {
+              display: block;
+              border: 0.1rem solid black;
+              border-radius: 0.25rem;
+              text-align: center;
+              margin: 0.5rem auto 0.15rem;
+              padding: 0.45rem;
+              width: 90vw;
+              background-color: $second-color;
+              font-size: 1.2rem;
+            }
 
-    .movies-nav {
-      display: block;
-      margin: 0 auto;
-    }
+            .nav {
+              width: 100%;
+              display: block;
 
-    .movies-nav-list {
-      .movies-nav-element {
-        width: 85vw;
-        background-color: $forth-color;
-        border-radius: 0.25rem;
-        margin-bottom: 0.15rem;
-      }
-    }
+               ul {
+                display: flex;
+              flex-direction: column;
+              justify-content: space-evenly;
+              margin: 0 auto;
+              
+              .movies-nav {
+                display: block;
+                margin: 0 auto;
+              }
 
-    .movies-nav-list:first-child:hover,
-    .movies-nav-list:first-child:focus {
-      .movies-nav-element {
-        display: block;
-        width: 85vw;
-      }
-    }
-  }
+              .movies-nav-list {
+                margin-top: 0.5rem;
+                width: 90vw;
+                .movies-nav-element {
+                  width: 90vw;
+                  background-color: $forth-color;
+                  border-radius: 0.25rem;
+                  margin-bottom: 0.15rem;
+                  &:not(:first-child) {
+                    margin-top: 0.4rem;
+                    background-color: $second-color;
+                  }
+                }
+              }
+
+              .movies-nav-list:first-child:hover,
+              .movies-nav-list:first-child:focus {
+                .movies-nav-element {
+                  display: block;
+                  width: 90vw;
+                }
+              }
+            }
+          }
 }
