+++ static/css/main.css
@@ -169,21 +169,25 @@ header nav menu .active {
   }
 }
 .nav {
+  display: block;
   width: 100%;
+  margin: 0 auto;
+}
+.nav ul {
   display: flex;
   flex-direction: row;
 }
-.nav .movies-nav {
+.nav ul .movies-nav {
   list-style-type: none;
   display: inline;
 }
-.nav .movies-nav-list {
+.nav ul .movies-nav-list {
   display: flex;
   flex-direction: column;
   justify-content: space-between;
   margin-top: 2vh;
 }
-.nav .movies-nav-list .movies-nav-element {
+.nav ul .movies-nav-list .movies-nav-element {
   width: fit-content;
   text-decoration: none;
   list-style-type: none;
@@ -196,17 +200,17 @@ header nav menu .active {
   position: relative;
   font-size: larger;
 }
-.nav .movies-nav-list .movies-nav-element a {
+.nav ul .movies-nav-list .movies-nav-element a {
   color: black;
 }
-.nav .movies-nav-list .movies-nav-element:not(:first-child) {
+.nav ul .movies-nav-list .movies-nav-element:not(:first-child) {
   display: none;
 }
-.nav .movies-nav-list:first-child:hover .movies-nav-element {
+.nav ul .movies-nav-list:first-child:hover .movies-nav-element {
   display: block;
   width: inherit;
 }
-.nav .movies-nav-list:first-child:hover .arrow-top {
+.nav ul .movies-nav-list:first-child:hover .arrow-top {
   display: none;
 }
 
@@ -215,25 +219,49 @@ header nav menu .active {
 }
 
 @media screen and (max-width: 640px) {
+  .mobile-nav-menu {
+    display: block;
+    border: 0.1rem solid black;
+    border-radius: 0.25rem;
+    text-align: center;
+    margin: 0.5rem auto 0.15rem;
+    padding: 0.45rem;
+    width: 90vw;
+    background-color: white;
+    font-size: 1.2rem;
+  }
   .nav {
     width: 100%;
+    display: block;
+  }
+  .nav ul {
     display: flex;
     flex-direction: column;
+    justify-content: space-evenly;
+    margin: 0 auto;
   }
-  .nav .movies-nav {
+  .nav ul .movies-nav {
     display: block;
     margin: 0 auto;
   }
-  .nav .movies-nav-list .movies-nav-element {
-    width: 85vw;
+  .nav ul .movies-nav-list {
+    margin-top: 0.5rem;
+    width: 90vw;
+  }
+  .nav ul .movies-nav-list .movies-nav-element {
+    width: 90vw;
     background-color: #d9d9d9;
     border-radius: 0.25rem;
     margin-bottom: 0.15rem;
   }
-  .nav .movies-nav-list:first-child:hover .movies-nav-element,
-  .nav .movies-nav-list:first-child:focus .movies-nav-element {
+  .nav ul .movies-nav-list .movies-nav-element:not(:first-child) {
+    margin-top: 0.4rem;
+    background-color: white;
+  }
+  .nav ul .movies-nav-list:first-child:hover .movies-nav-element,
+  .nav ul .movies-nav-list:first-child:focus .movies-nav-element {
     display: block;
-    width: 85vw;
+    width: 90vw;
   }
 }
 .accordion-cont {
