+++ under-construction.html
@@ -44,7 +44,11 @@
       </div>
     </header>
     <nav class="sec">
-      <ul class="nav sec-cont">
+      <button class="mobile-nav-menu">
+        Meny <span class="arrow-top" id="menu-span">&#9662;</span>
+      </button>
+      <div class="nav sec-cont">
+      <ul id="nav-id">
         <li class="movies-nav">
           <ul class="movies-nav-list">
             <li class="movies-nav-element">
@@ -105,6 +109,7 @@
           </ul>
         </li>
       </ul>
+    </div>
     </nav>
 
     <main class="sec">
